requirements.txt
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.0.5
faiss-cpu>=1.7.4
chromadb>=0.3.29
openai>=1.0.0
python-dotenv>=1.0.0
ipython>=8.0.0
jupyter>=1.0.0
tiktoken>=0.3.0
